+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added `isAscending`, `isDescending` getters and `fromString` method to `Direction`.
 - Added `NullHandling` enum.
+- Added default constructor, `isIgnoreCase`, `nullHandling`, `isAscending`, `isDescending`, `copyWith`, and `reverse` to `Order`.
+- Added `defaultDirection` constant to `Sort`.
 
 ## 0.2.0